requirements.txt
beautifulsoup4==4.4.1
click==6.6
cookies==2.2.1
Flask==0.11.1
haxor==0.3.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PySocks==1.5.7
python-forecastio==1.3.4
pytz==2016.4
requests==2.4.3
responses==0.5.1
simplejson==3.8.2
six==1.10.0
twilio==5.4.0
Werkzeug==0.11.10
wikipedia==1.4.0
yahoo-finance==1.2.1
